js2-mode: Initial support
This commit adds initial support for js2-mode.  This is not a 100%
complete representation but that is more due to the differences between
how Atom and Emacs theme in general and less about the work.  For the
most part this is straight forward but there is support for
face-remapping to allow for a more accurate theme.  This can be disabled
if necessary and is documented in the source.