Match Atom One dark syntax colors for line numbers
The original Atom One dark syntax color for "gutter" in hex is: #636D83

This can be checked here:

  https://github.com/atom/one-dark-syntax/blob/7197fc93a2f4d89169cf11dca36fc3ac63b053ef/styles/colors.less#L29

and also by inspecting the Atom editor in dev mode. Atom uses (described
in terms of this theme's variables) "atom-one-dark-gutter" for the
foreground color of line numbers, and "atom-one-dark-fg" for the
foreground color of the current line number.

Fixed:
- "atom-one-dark-accent" color was defined twice (keep original)